+++ packages/svelte/CHANGELOG.md
@@ -1,4 +1,4 @@
-# svelte-social-link
+# svelte-social-links
 
 ## 1.0.0
 
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies [[`1fbcb98`](https://github.com/Ennoriel/social-links/commit/1fbcb98183a1c9db69e3f58c2e736dec5612569a)]:
-  - static-social-link@1.0.0
+  - static-social-links@1.0.0
 
 ## 0.1.0
 
@@ -20,4 +20,4 @@
 ### Patch Changes
 
 - Updated dependencies []:
-  - static-social-link@0.1.0
+  - static-social-links@0.1.0